pyproject.toml
[project]
name = "procon-tester"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pygame-ce>=2.5.6",
    "pyglet>=2.1.13",
]
